refactor(client): move tree graph tab content to the separate component
